// src/views/personal/faka/product/utils/types.ts
// 虽然字段很少 但是抽离出来 后续有扩展字段需求就很方便了
interface MetaDTO {
  exchange?: number | string;
  skeys?: string;
  own_price?: number | string;
}
interface FormItemProps {
  id?: number;
  /** 用于判断是`新增`还是`修改` */
  title: string;
  name: string;
  category: string;
  price: number;
  inventory: string | number;
  sold?: string | number;
  description: string;
  data?: MetaDTO;
  skeys?: string;
  exchange?: number | string;
  own_price?: any;
  enable: string;
  CategoryOptions?: any[];
}
interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };